fix: المكافأةُ تُصرف · الطلبُ الخاصُّ يُوثَّق · وستُّ ملاحظاتٍ من الميدان
(دفعةُ المالك ٢٠٢٦-٠٨-١٨/١٩.)

## ١ · الهديّةُ لم تُصرف لأحدٍ منذ بُنيت

**ثلاثُ طبقاتٍ فوق بعضها، كشفها اختبارٌ ينفّذ الدالّةَ على قاعدةٍ
حقيقيّة:**

**أ · الوسيطُ نفسُه لعمودين من نوعين** — `user_id` من نوع uuid و`ref`
نصّ، وكلاهما `$1`. **ولا يكفي أن يُصبّ أحدُهما**: بوستغرس تستنتج نوعاً
واحداً من أوّل موضعٍ **فينقلب الخطأُ إلى الطرف الآخر.** فيُمرَّر
المعرّفُ وسيطين.

**ب · وكلُّ الأخطاء كانت مبتلعة** — `err != nil || already` في سطرٍ
واحد: «سقط الاستعلام» و«صُرفت سلفاً» يخرجان بالطريقة نفسِها. **وثلاثةُ
مواضعَ أخرى تخرج صامتة.** فأُنطقت كلُّها، **وأوّلُ تشغيلةٍ بعد النطق
قالت السببَ الحقيقيَّ فورا.**

**ج · وكشف الاختبارُ خطراً لم يُشتكَ منه**: القاعدةُ تمنع الرصيدَ
السالبَ إلّا لمحفظةٍ موسومةٍ خزينة. **ومنصّةٌ جديدةٌ لم تُوسَم خزينتُها
تبتلع كلَّ هديّةٍ بصمت.**

**والدرس**: أُبلغ المالكُ أنّها «مكتملة» بعد قراءة جسدها. **والقراءةُ
ليست إثباتا.**

## ٢ · توثيقُ سعر الطلب الخاصّ

`total = $2 + $3` — وسيطان مجهولان، **فتسأل بوستغرس: أيُّ «+» هذا؟**
فتسقط بخمسمئة، **والشاشةُ تترجمها «تعذّر الاتصال»** فيُتّهم الإنترنت.
ولم يكن على الدالّة اختبارٌ واحد.

## ٣ · الصورُ تتأخّر

**قِيس**: بطاقةُ الصنف تجلب الكاملةَ (٢٦٣ ك.ب) والمصغّرةُ ٤٠ —
**ستّةُ أضعافٍ ونصفٌ في بطاقةٍ عرضُها ١٢٣ نقطة**، وستُّ بطاقاتٍ في
الشاشة.

## ٤–٧ · الباقي

**عناوينُ الحساب تُخفى عن السائق** — العنوانُ حاجةُ من يُوصَّل إليه.
**وقسمٌ يُفتح فلا يُملأ أبداً يُقرأ نقصاً في الحساب.**

**وحذفُ صنفٍ من السلّة بضغطة** وإفراغُها — **والإنقاصُ حتّى الصفر
يجعل من أضاف خمسةً يضغط خمسا.**

**وأزرارُ العنوان في سطرين** — «اجعله العنوان الافتراضي» يأخذ سطرَه
كاملاً: **فعلٌ من نوعٍ آخرَ لا يزاحم التعديلَ والحذف.**

**والحسمُ يُطرح من الإجماليّ فورَ تطبيق الكود** — **وكان يُعرض سطراً
ولا يمسّ الرقم**، فيقرأ «خصم ٥٠٠» ويرى الإجماليَّ كما هو.

**وحديثُ السائق من بطاقة الطلب** — والمحرّكُ يسمح للطرفين منذ بُني،
**والزبونُ لم يملك باباً**: «دردشاتي السابقة» تُبنى من الرسائل فلا تعرض
طلباً لم يُكتب فيه شيء.

+++ backend/internal/orders/custom.go
@@ -130,6 +130,22 @@ func (s *Service) CreateCustom(ctx context.Context, customerID, request,
 //
 // **الاتّفاقُ قد يتبدّل**: يجد السائقُ الصنفَ أغلى فيعود إلى الزبون. **فيُحدَّث
 // ووقتُه معه** — والوقتُ هو ما يُقرأ عند الخلاف: **متى قال ماذا.**
+// # ولماذا يُكتب النوعُ في الجمع
+//
+// (شكوى المالك ٢٠٢٦-٠٨-١٨: «عند توثيق السعر بالطلب الخاصّ يعطي: تعذّر
+//
+//	الاتصال، حاول بعد قليل».)
+//
+// **وكان الجمعُ بلا نوع** — وسيطان مجهولان، **فتسأل بوستغرس: أيُّ «+»
+// هذا؟** فلا تجد واحداً بعينه ويسقط النداءُ بخمسمئة
+// (operator is not unique: unknown + unknown — رُئي في سجلّ الخادم).
+//
+// **والإسنادُ وحدَه يُعطي النوع** — والجمعُ لا: يُحسب قبل أن يُعرف
+// طرفاه.
+//
+// **ولا خطأَ يفهمه صاحبُ التطبيق**: تصل خمسُمئةٍ فتقول الشاشةُ «تعذّر
+// الاتصال» — **فيُتّهم الإنترنتُ وتُعاد المحاولةُ عشراً**، والعطبُ في
+// حرفين.
 func (s *Service) AgreeCustom(ctx context.Context, orderID, driverID string,
 	goods, fee int64) error {
 	if goods < 0 || fee < 0 {
@@ -186,7 +202,8 @@ func (s *Service) AgreeCustom(ctx context.Context, orderID, driverID string,
 	_, err = s.db.Exec(ctx, `
 		UPDATE orders
 		SET custom_goods_amount = $2, custom_fee = $3,
-		    subtotal = $2, delivery_fee = $3, total = $2 + $3,
+		    -- **والنوعُ يُقال صراحةً في الجمع** — انظر الشرحَ فوق الدالّة.
+		    subtotal = $2, delivery_fee = $3, total = $2::bigint + $3::bigint,
 		    custom_agreed_at = now(), updated_at = now()
 		WHERE id = $1`, orderID, goods, fee)
 	if err != nil {
